refactor(tests): mirror namespace layout in tests/Unit/
Move 20 test files into Spec/, Coverage/, and PHPUnit/ subdirectories so the
test directory matches the production layout introduced in the previous three
commits. Test class namespaces updated to Tests\Unit\{Spec,Coverage,PHPUnit}.

__DIR__-relative fixture paths (62 occurrences across 7 files) gained an
extra '/../' to compensate for the deeper directory level.

Spec/ (9 files):
  OpenApiSpecLoader, OpenApiSpecResolver, OpenApiSpecResolverFallback,
  OpenApiPathMatcher, OpenApiPathSuggester, OpenApiRefResolver,
  OpenApiRefResolverExternalRefs, OpenApiRefResolverHttpRefs,
  OpenApiSchemaConverter

Coverage/ (7 files):
  ConsoleCoverageRenderer, CoverageMergeCommand, CoverageSidecar,
  CoverageThresholdEvaluator, MarkdownCoverageRenderer,
  OpenApiCoverageTracker, OpenApiCoverageTrackerStateSerialization

PHPUnit/ (4 files):
  ConsoleOutput, CoverageReportSubscriberThreshold,
  CoverageReportSubscriberWorkerMode, OpenApiCoverageExtension

Top-level entry-point tests (OpenApiResponseValidator, OpenApiRequestValidator,
ValidatesOpenApiSchema*, etc.) stay flat in tests/Unit/.

PHPUnit 1022/1022 passing. PHPStan level 6 clean. PHP-CS-Fixer clean.
